+++ R/RcppExports.R
@@ -33,6 +33,10 @@ degree_tie <- function(type, edgelist, adjmat, actors, types, start, stop, consi
     .Call(`_remstats_degree_tie`, type, edgelist, adjmat, actors, types, start, stop, consider_type)
 }
 
+degree_undirected_tie <- function(type, edgelist, adjmat, actors, types, start, stop, consider_type) {
+    .Call(`_remstats_degree_undirected_tie`, type, edgelist, adjmat, actors, types, start, stop, consider_type)
+}
+
 triad_tie <- function(type, edgelist, actors, types, adjmat, riskset, start, stop, consider_type) {
     .Call(`_remstats_triad_tie`, type, edgelist, actors, types, adjmat, riskset, start, stop, consider_type)
 }
@@ -65,8 +69,12 @@ FEtype_tie <- function(covariates, edgelist, riskset, start, stop) {
     .Call(`_remstats_FEtype_tie`, covariates, edgelist, riskset, start, stop)
 }
 
-compute_stats_tie <- function(effects, edgelist, adjmat, actors, types, riskset, scaling, covariates, interactions, start, stop, directed, verbose) {
-    .Call(`_remstats_compute_stats_tie`, effects, edgelist, adjmat, actors, types, riskset, scaling, covariates, interactions, start, stop, directed, verbose)
+current_common_partners <- function(edgelist, riskset, actors, duration, start, stop) {
+    .Call(`_remstats_current_common_partners`, edgelist, riskset, actors, duration, start, stop)
+}
+
+compute_stats_tie <- function(effects, edgelist, adjmat, actors, types, riskset, scaling, covariates, interactions, start, stop, directed) {
+    .Call(`_remstats_compute_stats_tie`, effects, edgelist, adjmat, actors, types, riskset, scaling, covariates, interactions, start, stop, directed)
 }
 
 actorStat_rc <- function(covariates, edgelist, riskset, actors, start, stop, scaling) {
@@ -105,12 +113,16 @@ rrank_choice <- function(type, edgelist, riskset, actors, start, stop) {
     .Call(`_remstats_rrank_choice`, type, edgelist, riskset, actors, start, stop)
 }
 
-compute_stats_rate <- function(effects, edgelist, riskset, adjmat, actors, scaling, covariates, interactions, start, stop, verbose) {
-    .Call(`_remstats_compute_stats_rate`, effects, edgelist, riskset, adjmat, actors, scaling, covariates, interactions, start, stop, verbose)
+compute_stats_rate <- function(effects, edgelist, riskset, adjmat, actors, scaling, covariates, interactions, start, stop) {
+    .Call(`_remstats_compute_stats_rate`, effects, edgelist, riskset, adjmat, actors, scaling, covariates, interactions, start, stop)
 }
 
-compute_stats_choice <- function(effects, edgelist, adjmat, actors, riskset, scaling, covariates, interactions, start, stop, verbose) {
-    .Call(`_remstats_compute_stats_choice`, effects, edgelist, adjmat, actors, riskset, scaling, covariates, interactions, start, stop, verbose)
+compute_stats_choice <- function(effects, edgelist, adjmat, actors, riskset, scaling, covariates, interactions, start, stop) {
+    .Call(`_remstats_compute_stats_choice`, effects, edgelist, adjmat, actors, riskset, scaling, covariates, interactions, start, stop)
+}
+
+compute_adjmat_update <- function(edgelist, adjmat, N, D, directed, memory, memory_value, start, stop) {
+    .Call(`_remstats_compute_adjmat_update`, edgelist, adjmat, N, D, directed, memory, memory_value, start, stop)
 }
 
 # Register entry points for exported C++ functions